% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reading.R
\name{brom_id_maker}
\alias{brom_id_maker}
\title{Make id column}
\usage{
brom_id_maker(df)
}
\arguments{
\item{df}{the data.frame. must contain site and
  bromelaid.id columns}
}
\value{
the data frame, plus a new column for the id
}
\description{
This joins the \code{site} and the \code{bromeliad.id}
columns. to make a unique identifier.
}

