% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sheetname_read.R
\name{bromeliad.final.inverts_read}
\alias{bromeliad.final.inverts_read}
\title{Read in the final.inverts tab}
\usage{
bromeliad.final.inverts_read(file_to_read)
}
\arguments{
\item{file_to_read}{Path to file to be read}
}
\description{
this function reads one final.inverts sheet
}

