% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheetname_read.R
\name{bromeliad.ibuttons_read}
\alias{bromeliad.ibuttons_read}
\title{Read in the bromeliad.ibuttons tab}
\usage{
bromeliad.ibuttons_read(file_to_read)
}
\arguments{
\item{file_to_read}{Path to file to be read}
}
\description{
this function reads one site.info sheet
}

