% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sheetname_read.R
\name{bromeliad.physical_read}
\alias{bromeliad.physical_read}
\title{Read in the bromeliad.physical tab}
\usage{
bromeliad.physical_read(file_to_read)
}
\arguments{
\item{file_to_read}{Path to file to be read}
}
\description{
this function reads bromeliad.physical sheet
}

