% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheetname_read.R
\name{bromeliad.terrestrial_read}
\alias{bromeliad.terrestrial_read}
\title{Read in the bromeliad.terrestrial tab}
\usage{
bromeliad.terrestrial_read(file_to_read)
}
\arguments{
\item{file_to_read}{Path to file to be read}
}
\description{
this function reads bromeliad.terrestrial sheet
}

