% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-bwgtools.R
\docType{package}
\name{bwgtools}
\alias{bwgtools}
\alias{bwgtools-package}
\title{bwgtools: A package for the bromeliad working group: the rainfall experiment}
\description{
The bwgtools package provides four categories of important functions:
Accessing data, Combining data, Calulating metrics, and Analyzing data.
}
\section{Accessing data}{

\itemize{
\item \code{\link{offline}}
}
}

\section{Combining data}{

\itemize{
\item \code{\link{combine_tab}}
\item \code{\link{read_site_sheet}}
}
}

\section{Calculating metrics}{

\itemize{Hydrological metrics
\item \code{\link{hydro_variables}}
\item \code{\link{water_summary_calc}}
}
\itemize{Rainfall metrics
}
\itemize{Functional groups and trophic levels
\item \code{\link{sum_func_groups}}
}
\itemize{Decomposition
\item \code{\link{get_decomp}}
}
}

\section{Analyzing data}{

\itemize{
\item \code{\link{make_rlq}}
}
}
\seealso{
\url{https://github.com/SrivastavaLab/bwgtools/blob/master/README.Rmd}
}

