% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checking.R
\name{check_names}
\alias{check_names}
\title{Checks names of a dataset}
\usage{
check_names(dataset, column_names = c("site", "trt.name", "bromeliad.id",
  "date", "depth.centre.measure.first", "depth.leafa.measure.first",
  "depth.leafb.measure.first", "depth.centre.water.first",
  "depth.leafa.water.first", "depth.leafb.water.first"))
}
\arguments{
\item{dataset}{the dataset in question (as a data.frame)}

\item{column_names}{the correct names of that spreadsheet}
}
\description{
could possibly be adapted for all datasets
}

