% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combining.R
\name{combine_site.info}
\alias{combine_site.info}
\title{Combines all site.info tabs}
\usage{
combine_site.info()
}
\value{
data.frame containing all the site information
}
\description{
Combines all site.info tabs
}

