% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculations.R
\name{decomp_responses}
\alias{decomp_responses}
\title{Species decomp and total decomp}
\usage{
decomp_responses(leaf_loss_species)
}
\arguments{
\item{leaf_loss_species}{the data to combine. must be the output of \code{leaf_loss_mean()}}
}
\value{
a data.frame, one row per bromeliad
}
\description{
Species decomp and total decomp
}

