% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fixing.R
\name{fill_in_argentina}
\alias{fill_in_argentina}
\title{Fill in argentina data}
\usage{
fill_in_argentina(argentina_data)
}
\arguments{
\item{argentina_data}{the argentina dataset}
}
\value{
the same data, but with days filled over
}
\description{
This function needs TESTING
}

