% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checking.R
\name{filter_centre_leaf}
\alias{filter_centre_leaf}
\title{Filter out the centre leaf}
\usage{
filter_centre_leaf(data, centre_filter = TRUE)
}
\arguments{
\item{data}{dataset to filter. must contain a column called "leaf"}

\item{centre_filter}{do you want to drop the central leaf? defaults to TRUE}
}
\value{
data.frame without the centre leaf
}
\description{
Filter out the centre leaf
}

