% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterdepths_calculations.R
\name{filter_long_water}
\alias{filter_long_water}
\title{Filter water data (long format)}
\usage{
filter_long_water(Data, rm_centre = TRUE)
}
\arguments{
\item{Data}{long-format water data}

\item{rm_centre}{should the centre leaf be used? defaults to TRUE}
}
\value{
ungrouped data.frame
}
\description{
This function filters long-format water data, by applying \code{filter_centre_leaf}
 and \code{filter_naonly_groups}
}

