% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date_range.R
\name{firstday}
\alias{firstday}
\title{Find first days (for duration)}
\usage{
firstday(df)
}
\arguments{
\item{df}{the dataframe of water depths}
}
\value{
A dataframe with three columns: trt.name, bromeliad.id, startdate.
}
\description{
We want to get the first date for each bromeliad. This function only works if the dataframe contains the \code{date} and the \code{bromeliad.id} columns (ex: \code{sheetname="leaf.waterdepth"}).
}
\examples{
Arg <- read_site_sheet("Argentina","leaf.waterdepths")
firstday(Arg)
}

