% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combining.R
\name{get_all_insects}
\alias{get_all_insects}
\title{Get all the final insect data from Dropbox}
\usage{
get_all_insects(site_names = c("Macae", "PuertoRico", "French_Guiana"))
}
\arguments{
\item{site_names}{Character vector of sites with (correctly formatted!) column names}
}
\value{
data.frame with a column for "site"
}
\description{
Get all the final insect data from Dropbox
}

