% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{get_all_sites}
\alias{get_all_sites}
\title{Get all that data}
\usage{
get_all_sites(sites = c("Argentina", "Cardoso", "Colombia", "French_Guiana",
  "Macae", "PuertoRico", "CostaRica"), sheetname = NULL)
}
\arguments{
\item{sites}{a vector of all the sites you want}

\item{sheetname}{the sheet name}
}
\description{
Get all that data
}

