% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reading.R
\name{get_bwg_names}
\alias{get_bwg_names}
\title{obtain the full name data from github}
\usage{

  get_bwg_names(file = "https://raw.githubusercontent.com/SrivastavaLab/bwg_names/master/data/Distributions_organisms_full.tsv",
  chars = 23, nums = 49)
}
\arguments{
\item{file}{the location of the file. defaults to the internet}

\item{chars}{the number of character columns. defaults to 21}

\item{nums}{the number of numeric columns. defaults to 54}
}
\value{
A dataframe containing all information as in the excel file in Dropbox: "Distribution organisms_correct2015"
}
\description{
Gets the complete insect taxonomic data from github. This is good
because it will always be accurate.
Just type in \code{get_bwg_names()} in R's Console and you will get all the taxonomic information (among other things).
}
\examples{
taxo <- get_bwg_names()
names(taxo)
}

