% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterdepths_calculations.R
\name{group_or_summarize}
\alias{group_or_summarize}
\title{Group or summarize water data (long form)}
\usage{
group_or_summarize(data, aggregate_leaves = FALSE)
}
\arguments{
\item{data}{a data.frame created by \code{longwater()}}

\item{aggregate_leaves}{should leaves be aggregated? defaults to \code{FALSE}}
}
\value{
if \code{aggregate = FALSE}, a data.frame of the same size as the original, but with groups defined for further processing. If \code{aggregate_leaves = TRUE}, output has one observation per bromeliad (rather than one observation per leaf within bromeliad)
}
\description{
Group or summarize water data (long form)
}

