% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterdepths_calculations.R
\name{hydro_variables}
\alias{hydro_variables}
\title{Calculate the hydrological variables}
\usage{
hydro_variables(waterdata, sitedata, physicaldata, rm_centre = TRUE,
  aggregate_leaves = FALSE, .columns = "measure")
}
\arguments{
\item{waterdata}{the leaf.waterdepths tab}

\item{sitedata}{the site.info tab}

\item{physicaldata}{the bromeliad.physical}

\item{rm_centre}{remove centre? defaults to TRUE}

\item{aggregate_leaves}{aggregate leaves? defaults to FALSE}
}
\value{
The hydrological variables
}
\description{
Calculate the hydrological variables
}
\examples{
leafwater <- combine_tab(c("Argentina", "French_Guiana"),"leaf.waterdepths")
sites <- combine_tab(c("Argentina", "French_Guiana"),"site.info")
phys <- combine_tab(c("Argentina", "French_Guiana"),"bromeliad.physical")
hydro <- hydro_variables(waterdata = leafwater,
                          sitedata = sites,
                      physicaldata = phys)
hydro
}

