% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining.R
\name{invert_to_long}
\alias{invert_to_long}
\title{Tidy wide invert data into long format}
\usage{
invert_to_long(insect_data, category_vars)
}
\arguments{
\item{insect_data}{the dataset to tidy. In the same shape as the bromeliad.inverts.final tab}

\item{category_vars}{those variables which define groups (ie the names of every variable that is NOT the name of an invertebrate species). Must be a character vector.}
}
\value{
tbl.df containing invertebrate data in long format
}
\description{
Tidy wide invert data into long format
}

