% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sheetname_read.R
\name{leaf.waterdepths_read}
\alias{leaf.waterdepths_read}
\title{Read in the leaf.waterdepths tab}
\usage{
leaf.waterdepths_read(file_to_read)
}
\arguments{
\item{file_to_read}{Path to file to be read}
}
\description{
this function reads one water depth sheet
}

