% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculations.R
\name{leaf_loss_mean}
\alias{leaf_loss_mean}
\title{Calculate average decomposition for each species}
\usage{
leaf_loss_mean(leaf_loss_sample_data)
}
\arguments{
\item{leaf_loss_sample_data}{the sample loss data}
}
\value{
means and sample sizes for each species
}
\description{
Calculate average decomposition for each species
}

