% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decomposition.R
\name{leaf_loss_sample}
\alias{leaf_loss_sample}
\title{calculate loss for each sample}
\usage{
leaf_loss_sample(long_phys_data)
}
\arguments{
\item{long_phys_data}{physical data in long format}
}
\value{
adds column "loss"
}
\description{
calculate loss for each sample
}

