% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/waterdepths_calculations.R
\name{longwater}
\alias{longwater}
\title{make the water data long}
\usage{
longwater(df, which_columns = "measure")
}
\arguments{
\item{df}{the water data to make long}
}
\value{
long water data
}
\description{
make the water data long
}

