% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/waterdepths_calculations.R
\name{make_full_timeline}
\alias{make_full_timeline}
\title{fill in the missing timeline}
\usage{
make_full_timeline(filtered_water_data, sitedata, physdata)
}
\arguments{
\item{filtered_water_data}{a pre-filtered water dataset. acts as a "template" for the grouping variables to include}

\item{sitedata}{contents of the site.info tab}

\item{physdata}{contents of the physical.info tab}
}
\value{
a long dataset that contains a consecutive series of dates
}
\description{
This function fills in the complete sequence of dates,
from the beginning to the end of a site.
}

