% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RLQ.R
\name{make_matrix}
\alias{make_matrix}
\title{Create a matrix from a data.frame}
\usage{
make_matrix(df, rownm = "species")
}
\arguments{
\item{df}{a dataframe. should be all numeric except for
the first column, which gives rownames.}

\item{rownm}{name of the first column, which is to become rownames}
}
\value{
a numeric matrix
}
\description{
In order to make a matrix for RLQ analysis, we need to
convert data.frames containing data into a matrix format.
In these data.frames the first column holds what should
be the rownames of the matrix.
}

