% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/waterdepths_calculations.R
\name{make_support_file}
\alias{make_support_file}
\title{Make the support file}
\usage{
make_support_file(allsites, phys)
}
\arguments{
\item{allsites}{combined site.info tab}

\item{phys}{combined bromeliad.physical tab}
}
\value{
A dataframe containing support file: site.name, trt.name, temporal.block, start_block and finish_block. Start_block is equivalent to start.water.addition from the site.info tab, and finish_block is equivalent to last.day.sample
}
\description{
Make the support file
}
\examples{
sites <- combine_tab(c("Argentina", "French_Guiana"),"site.info")
phys <- combine_tab(c("Argentina", "French_Guiana"),"bromeliad.physical")
sup_file <- make_support_file(sites,phys)
sup_file
}

