% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combining.R
\name{merge_func}
\alias{merge_func}
\title{Merge functional groups to invert data}
\usage{
merge_func(insect_data, trait_data)
}
\arguments{
\item{insect_data}{data.frame of invert observations. must be long format (ie output of \code{invert_to_long})}

\item{trait_data}{bwg_names data. output of \code{get_bwg_names}}
}
\value{
merged data
}
\description{
Merge functional groups to invert data
}

