% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checking.R
\name{names_all_same}
\alias{names_all_same}
\title{Check that names are all identical}
\usage{
names_all_same(datalist)
}
\arguments{
\item{datalist}{list of dataframes whose names must all be identical}
}
\value{
are the names identical? TRUE or FALSE
}
\description{
Check that names are all identical
}

