% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sheetname_read.R
\name{neutral_read}
\alias{neutral_read}
\title{read in a neutral file}
\usage{
neutral_read(sheetname)
}
\arguments{
\item{sheetname}{name of the sheet you're reading}
}
\value{
a function that reads that sheet with col_types = NULL
}
\description{
read in a neutral file
}

