% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reading.R
\name{offline}
\alias{offline}
\title{Get a file while offline}
\usage{
offline(sitename = c("Argentina", "Cardoso", "Colombia", "French_Guiana",
  "Macae", "PuertoRico", "CostaRica"), default.path = "../../")
}
\arguments{
\item{sitename}{Name of a single site you wish to read}

\item{default.path}{the default path to your Dropbox folder. Defaults to one that works within a paper folder, assuming you have not rearranged your folders within dropbox.}
}
\value{
a correct relative path
}
\description{
If you don't have access to dropbox, you can still read files from your local dropbox folder. This function creates the path to the files.
}

