% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decomposition.R
\name{physical_long}
\alias{physical_long}
\title{Make physical data long}
\usage{
physical_long(physical_data)
}
\arguments{
\item{physical_data}{physical data, as produced by \code{combine_tab()}}
}
\value{
data.frame
}
\description{
Rearranges the contents of the bromeliad.physical tab into a long format
}

