% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combining.R
\name{plot_trophic}
\alias{plot_trophic}
\title{Plot predator vs prey biomass}
\usage{
plot_trophic(invert_data, trait_data)
}
\arguments{
\item{trait_data}{bwg_names data. output of \code{get_bwg_names}}

\item{insect_data}{data.frame of invert observations. must be long format (ie output of \code{invert.long})}
}
\value{
a ggplot
}
\description{
Plot predator vs prey biomass
}

