% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/demo.R
\name{read_site_sheet}
\alias{read_site_sheet}
\title{read in a sheet from all sites}
\usage{
read_site_sheet(sitename, sheetname = NULL)
}
\arguments{
\item{sheetname}{the sheet name you want. must match exactly.}

\item{file}{the file name to fetch}
}
\description{
this function reads all the sheets in an excel book on dropbox
}

