% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sheetname_read.R
\name{site.info_read}
\alias{site.info_read}
\title{Read in the site.info tab}
\usage{
site.info_read(file_to_read)
}
\arguments{
\item{file_to_read}{Path to file to be read}
}
\description{
this function reads one site.info sheet
}

