% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combining.R
\name{sum_trophic}
\alias{sum_trophic}
\title{Summarize functional groups still farther into trophic ranks}
\usage{
sum_trophic(func_sums)
}
\arguments{
\item{func_sums}{must be a grouped tbl.df, the groups must be bromeliad.id and pred_prey, in that order}
}
\value{
summarized data
}
\description{
Summarize functional groups still farther into trophic ranks
}

