% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheetname_read.R
\name{terrestrial.taxa_read}
\alias{terrestrial.taxa_read}
\title{Read in the terrestrial.taxa tab}
\usage{
terrestrial.taxa_read(file_to_read)
}
\arguments{
\item{file_to_read}{Path to file to be read}
}
\description{
this function reads terrestrial.taxa sheet
}

