% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/waterdepths_calculations.R
\name{water_summary_calc}
\alias{water_summary_calc}
\title{Calculate water depth measurements}
\usage{
water_summary_calc(depth)
}
\arguments{
\item{depth}{depth measurements}
}
\value{
a 1 x n row \code{tbl_df}
}
\description{
Calculate water depth measurements
}
\details{
text describing parameter inputs in more detail.
\itemize{
 \item{"max.depth"}{ the maximum depth}
 \item{"min.depth"}{ the minimum depth}
 \item{"mean.depth"}{ mean depth}
 \item{"var.depth"}{ variance in depth}
 \item{"sd.depth"}{ standard deviation in depth}
 \item{"cv.depth"}{ coefficient of variation in depth}
 \item{"net_fluct"}{ net fluctuation in depth}
 \item{"total_fluct"}{ total fluctuation in depth}
 \item{"amplitude"}{ max.depth - mean.depth}
 \item{"wetness"}{mean.depth / max.depth}
}
}

