% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining.R
\name{which_names_doubled}
\alias{which_names_doubled}
\title{Find and edit duplicate names in dataset}
\usage{
which_names_doubled(df)
}
\arguments{
\item{df}{data frame which might have duplicate names}
}
\value{
data.frame with unique names. if the names were already unique, it is the same. otherwise the names are passed through \code{make.names} and a message is given
}
\description{
Find and edit duplicate names in dataset
}

