% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package-bwgtools.R
\docType{package}
\name{bwgtools}
\alias{bwgtools}
\alias{bwgtools-package}
\title{bwgtools: A package for the bromeliad working group: the rainfall experiment}
\description{
The bwgtools package provides four categories of important functions:
Accessing data, Combining data, Calulating metrics, and Analyze data.
}
\section{Accessing data}{

\code{\link{offline}}
}

\section{Combining data}{

\code{\link{combine_tab}}
}

\section{Calculating metrics}{

\code{\link{hydro_variable}}
}

\section{Analyzing data}{

\code{\link{make_rlq}}
}

