% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combining.R
\name{combine_tab}
\alias{combine_tab}
\title{Obtain data for all sites}
\usage{
combine_tab(.sites = c("Argentina", "Cardoso", "Colombia", "French_Guiana",
  "Macae", "PuertoRico", "CostaRica"), sheetname)
}
\arguments{
\item{.sites}{the sites you want. defaults to all of them}

\item{sheetname}{The name of the sheet you'd like to have}
}
\value{
data.frame of all bromeliad.physical tabs
}
\description{
This function reads data from the same tab for all the sites (via \code{get_all_sites()}) , then combines thme with \code{dplyr::rbind_all()}
}

