% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checking.R
\name{filter_naonly_groups}
\alias{filter_naonly_groups}
\title{Filter out any groups where depth is all NA}
\usage{
filter_naonly_groups(data)
}
\arguments{
\item{data}{a grouped data.frame}
}
\value{
a data.frame with all NA groups gone
}
\description{
Filter out any groups where depth is all NA
}

