% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combining.R
\name{find_site_brom}
\alias{find_site_brom}
\title{Are the columns site and bromeliad.id found in this data.frame?}
\usage{
find_site_brom(df)
}
\arguments{
\item{df}{data frame to check for}
}
\value{
are those column names present? TRUE or FALSE
}
\description{
Are the columns site and bromeliad.id found in this data.frame?
}

