% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hydrological.R
\name{firstday}
\alias{firstday}
\title{Find first days (for duration)}
\usage{
firstday(df)
}
\arguments{
\item{df}{the dataframe of water depths}
}
\description{
We want to get the first date for each bromeliad
}

