% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date_range.R
\name{from_start}
\alias{from_start}
\title{Difference in days}
\usage{
from_start(depthdata)
}
\arguments{
\item{df}{a dataframe containing \code{sheetname="leaf.waterdepths"}}
}
\value{
A dataframe combining the \code{date} and the \code{bromeliad.id}, plus a new column (\code{nday}) for the number of days since the begining of the experiment.
}
\description{
calculate the difference in days to the start of the experiment
}
\examples{
Arg <- read_site_sheet("Argentina","leaf.waterdepths")
from_start(Arg)
}

