% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/date_range.R
\name{from_start}
\alias{from_start}
\title{Difference in days}
\usage{
from_start(depthdata)
}
\arguments{
\item{df}{the leaf.waterdepths data.frame}
}
\description{
calculate the difference in days to the start of the experiment
}

