% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/physical_calculations.R
\name{get_decomp}
\alias{get_decomp}
\title{obtain decomposition data}
\usage{
get_decomp()
}
\value{
data frame with 7 columns: site, trt.name, bromeliad.id, species1 decomposition, species2 decomposition, mean decomposition, and the number of species contained in that mean
}
\description{
obtain decomposition data
}

