% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reading.R
\name{leaf.waterdepths_read}
\alias{leaf.waterdepths_read}
\title{Read in the leaf.waterdepths tab}
\usage{
leaf.waterdepths_read(file_to_read)
}
\description{
this function reads one water depth sheet
}

