% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reading.R
\name{make_default_path}
\alias{make_default_path}
\title{Make default dropbox path}
\usage{
make_default_path(.sitename)
}
\arguments{
\item{.sitename}{quoted site name}
}
\value{
the default path to file from dropbox
}
\description{
Make default dropbox path
}

