% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/demo.R
\name{read_sheet}
\alias{read_sheet}
\title{Read an excel file from dropbox account}
\usage{
read_sheet(file, dest = tempdir(), sheetname, ...)
}
\description{
this function reads all the sheets in an excel book on dropbox
}

