% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reading.R
\name{read_site_sheet}
\alias{read_site_sheet}
\title{read in a sheet from all sites}
\usage{
read_site_sheet(sitename, sheetname = NULL, ...)
}
\arguments{
\item{sitename}{The name of the site you want to read data from, or a path to where the excel files are on your local computer (for offline work):\code{"Argentina"}, \code{"Cardoso"}, \code{"Colombia"}, \code{"French_Guiana"}, \code{"Macae"}, \code{"PuertoRico"}, \code{"CostaRica"}}

\item{sheetname}{The name of the sheetname you want: \code{"leaf.waterdepths"}, \code{"bromeliad.physical"}, \code{"bromeliad.final.inverts"}, \code{"site.info"}, \code{"site.weather"}, \code{"bromeliad.initial.inverts"}, \code{"bromeliad.terrestrial"}, \code{"terrestrial.taxa"}}

\item{file}{The file name to fetch}

\item{OnDisk}{Are you reading from a local copy or Dropbox? defaults to FALSE}
}
\description{
This function reads all the sheets in an excel book on dropbox
}
\examples{
macae <- read_site_sheet(("Macae"), "leaf.waterdepths")
}

